requirements.txt
python-dateutil>=2.5.2,<3.0.0
requests>=2.9.1,<3.0.0
meza>=0.31.1,<0.32.0
manage.py~=0.2.10
